Separate the code required for the support of different MPI implementations
